requirements.txt
streamlit
smtplib
requests
email
python-dotenv
